openreservation/requirements.txt
apipkg==1.5
asgiref==3.2.10
atomicwrites==1.4.0; sys_platform == "win32"
attrs==19.3.0
babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3; sys_platform == "win32"
coverage==5.1
defusedxml==0.6.0
django==3.0.7
django-allauth==0.41.0
django-mkdocs==0.1.2
django-phonenumber-field==4.0.0
django-pytest==0.2.0
execnet==1.7.1
future==0.18.2
idna==2.9
importlib-metadata==1.6.1; python_version < "3.8"
jinja2==2.11.2
joblib==0.15.1; python_version > "2.7"
livereload==2.6.2
lunr==0.5.8
markdown==3.2.2
markupsafe==1.1.1
mkdocs==1.1.2
more-itertools==8.4.0
nltk==3.5; python_version > "2.7"
oauthlib==3.1.0
packaging==20.4
phonenumberslite==8.12.6
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-forked==1.1.3
pytest-xdist==1.32.0
python3-openid==3.1.0
pytz==2020.1
pyyaml==5.3.1
regex==2020.6.8; python_version > "2.7"
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1; python_version > "2.7"
urllib3==1.25.9
wcwidth==0.2.4
whitenoise==5.1.0
zipp==3.1.0; python_version < "3.8"
